Allow injecting variables to Products component
Having properties extendable is great, but it's not much use if there's
no way of assigning them to variables so they can be used from the
component class and any partials it loads. This adds a crude way of
enabling other plugins to add their own variables via value or callback.